% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{irGSEA.barplot}
\alias{irGSEA.barplot}
\title{Stacked bar plot}
\usage{
irGSEA.barplot(
  object = NULL,
  method = NULL,
  significance.color = NULL,
  color.cluster = NULL,
  color.method = NULL,
  cluster.levels = NULL
)
}
\arguments{
\item{object}{A list after perform \code{\link{irGSEA.integrate}}}

\item{method}{A vector. It should be one or more of the followling : AUCell,
UCell, singscore, ssgsea or RRA. Default all methods when it is set to NULL.}

\item{significance.color}{A vector. Default "c("#D0DFE6FF","#f87669")" when
it is set to NULL.}

\item{color.cluster}{A vector. Default "ggsci::pal_igv()(the number of colnames
of enrichment score matrix)" when it is set to NULL.}

\item{color.method}{A vector. Default "ggsci::pal_igv()(the number of methods
)" when it is set to NULL.}

\item{cluster.levels}{A vector equal to the number of clusters.}
}
\value{
stacked bar plot
}
\description{
Easy to show analysis results by stacked bar plot
}
\examples{
\dontrun{
# load PBMC dataset by R package SeuratData
library(Seurat)
library(SeuratData)
# download 3k PBMCs from 10X Genomics
InstallData("pbmc3k")
data("pbmc3k.final")
pbmc3k.final <- SeuratObject::UpdateSeuratObject(pbmc3k.final)

# Seurat object
pbmc3k.final <- irGSEA.score(object = pbmc3k.final, assay = "RNA",
slot = "data", msigdb = T, species = "Homo sapiens",
category = "H", geneid = "symbol",
method = c("AUCell", "UCell", "singscore", "ssgsea"), kcdf = 'Gaussian')

# Integrated analysis
result.dge <- irGSEA.integrate(object = pbmc3k.final,
group.by = "seurat_annotations", metadata = NULL, col.name = NULL,
method = c("AUCell","UCell","singscore","ssgsea"))

irGSEA.barplot.plot1 <- irGSEA.barplot(object = result.dge)

}

}
