% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irGSEA.density.scatterplot.R
\name{irGSEA.density.scatterplot}
\alias{irGSEA.density.scatterplot}
\title{Density Scatter plot}
\usage{
irGSEA.density.scatterplot(
  object = NULL,
  method = NULL,
  show.geneset = NULL,
  reduction = "umap",
  ...
)
}
\arguments{
\item{object}{A Seurat after perform \code{\link{irGSEA.score}}}

\item{method}{A character. It should be one of the followling : AUCell,
UCell, singscore, ssgsea.}

\item{show.geneset}{A character. It should be one of the rownames of
enrichment score matrix.}

\item{reduction}{A character. It can not be empty and should be calculated
in advance.}

\item{...}{More parameters pass to \code{\link[Nebulosa]{plot_density}}}
}
\value{
density scatter plot
}
\description{
Easy to the data distribution by density scatter plot
}
\examples{
\dontrun{
# load PBMC dataset by R package SeuratData
library(Seurat)
library(SeuratData)
# download 3k PBMCs from 10X Genomics
InstallData("pbmc3k")
data("pbmc3k.final")
pbmc3k.final <- SeuratObject::UpdateSeuratObject(pbmc3k.final)

# Seurat object
pbmc3k.final <- irGSEA.score(object = pbmc3k.final, assay = "RNA",
slot = "data", msigdb = T, species = "Homo sapiens",
category = "H", geneid = "symbol",
method = c("AUCell", "UCell", "singscore", "ssgsea"), kcdf = 'Gaussian')

irGSEA.density.scatterplot1 <- irGSEA.density.scatterplot(object = pbmc3k.final,
method = "UCell", show.geneset = "HALLMARK-INFLAMMATORY-RESPONSE",
reduction = "umap")
irGSEA.density.scatterplot2 <- irGSEA.density.scatterplot(object = pbmc3k.final,
method = "ssgsea", show.geneset = "HALLMARK-IL6-JAK-STAT3-SIGNALING",
reduction = "umap")

}


}
