% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{irGSEA.heatmap}
\alias{irGSEA.heatmap}
\title{Heatmap plot}
\usage{
irGSEA.heatmap(
  object = NULL,
  method = "RRA",
  top = 50,
  show.geneset = NULL,
  cluster_rows = T,
  significance.color = NULL,
  cluster.color = NULL,
  direction.color = NULL,
  rowname.fointsize = 7,
  heatmap.width = 17,
  heatmap.heigh = 13,
  cluster.levels = NULL,
  ...
)
}
\arguments{
\item{object}{A list after perform \code{\link{irGSEA.integrate}}}

\item{method}{A character. It should be one of the followling : AUCell, UCell,
singscore, ssgsea or RRA. Default RRA.}

\item{top}{The top gene sets. Default 50.}

\item{show.geneset}{A vector including special gene sets. Default NULL.}

\item{cluster_rows}{Whether to make cluster on rows. Defaul True.}

\item{significance.color}{A vector. Default "c("#D0DFE6FF","#f87669")" when
it is set to NULL.}

\item{cluster.color}{A vector. Default "ggsci::pal_igv()(the number of colnames
of enrichment score matrix)" when it is set to NULL.}

\item{direction.color}{A vector. Default "c("#4575B4","#D73027")" when it
is set to NULL.}

\item{rowname.fointsize}{The fointsize of rownames. Default 7.}

\item{heatmap.width}{Width of the whole heatmap (including heatmap
components), default 17.}

\item{heatmap.heigh}{Height of the whole heatmap (including heatmap
components), default 13.}

\item{cluster.levels}{A vector equal to the number of clusters.}

\item{...}{More parameters pass to \code{\link[ComplexHeatmap]{Heatmap}}}
}
\value{
heatmap plot
}
\description{
Easy to show analysis results by heatmap plot
}
\examples{
\dontrun{
# load PBMC dataset by R package SeuratData
library(Seurat)
library(SeuratData)
# download 3k PBMCs from 10X Genomics
InstallData("pbmc3k")
data("pbmc3k.final")
pbmc3k.final <- SeuratObject::UpdateSeuratObject(pbmc3k.final)

# Seurat object
pbmc3k.final <- irGSEA.score(object = pbmc3k.final, assay = "RNA",
slot = "data", msigdb = T, species = "Homo sapiens",
category = "H", geneid = "symbol",
method = c("AUCell", "UCell", "singscore", "ssgsea"), kcdf = 'Gaussian')

# Integrated analysis
result.dge <- irGSEA.integrate(object = pbmc3k.final,
group.by = "seurat_annotations", metadata = NULL, col.name = NULL,
method = c("AUCell","UCell","singscore","ssgsea"))

irGSEA.heatmap.plot1 <- irGSEA.heatmap(object = result.dge, method = "RRA",
top = 50, show.geneset = NULL)

irGSEA.heatmap.plot2 <- irGSEA.heatmap(object = result.dge, method = "ssgsea",
top = 50, show.geneset = NULL)

}



}
