% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irGSEA.hub.R
\name{irGSEA.hub}
\alias{irGSEA.hub}
\title{Calculate the hub gene of the geneset}
\usage{
irGSEA.hub(
  object = NULL,
  assay = "RNA",
  slot = "data",
  method = NULL,
  show.geneset = NULL,
  ncores = 4,
  type = "rank",
  maxRank = 2000,
  top = 5,
  correlation.color = c("#0073c2", "white", "#efc000"),
  method.color = NULL
)
}
\arguments{
\item{object}{A Seurat after perform \code{\link{irGSEA.score}}}

\item{assay}{Name of assay to calculate the correction. Default RNA.}

\item{slot}{Default data.}

\item{method}{A character.}

\item{show.geneset}{A character.}

\item{ncores}{Default 4.}

\item{type}{expression or rank. Default rank. Calculate the correlation between
the geneset's score and the  rank of gene included in the geneset while the
type is rank. Calculate the correlation between the geneset's score and the
expression of gene included in the geneset while the type is expression.}

\item{maxRank}{Maximum number of genes to rank per cell; above this rank,
a given gene is considered as not expressed. Default 2000.}

\item{top}{The number of top-ranked positively correlated genes in each method
is displayed in a heatmap.}

\item{correlation.color}{A vector.}

\item{method.color}{A vector. Default "ggsci::pal_png()(the number of
methods)" when it is set to NULL.}
}
\value{
list includes hub_result and hub_plot
}
\description{
Easy to calculate the hub gene of the geneset based on the correlation between
the geneset's score and the expression or rank of gene included in the geneset
}
\examples{
\dontrun{
# load PBMC dataset by R package SeuratData
library(Seurat)
library(SeuratData)
# download 3k PBMCs from 10X Genomics
InstallData("pbmc3k")
data("pbmc3k.final")
pbmc3k.final <- SeuratObject::UpdateSeuratObject(pbmc3k.final)

# Seurat object
pbmc3k.final <- irGSEA.score(object = pbmc3k.final, assay = "RNA",
slot = "data", msigdb = T, species = "Homo sapiens",
category = "H", geneid = "symbol",
method = c("AUCell", "UCell", "singscore", "ssgsea"), kcdf = 'Gaussian')

hub.result <- irGSEA.hub(object = pbmc3k.final, assay = "RNA", slot = "data",
method = c("AUCell","UCell","singscore", "ssgsea"),
show.geneset = c("HALLMARK-INFLAMMATORY-RESPONSE", "HALLMARK-APOPTOSIS"),
ncores = 4, type = "rank", maxRank = 2000, top = 5,
correlation.color = c("#0073c2","white","#efc000"), method.color = NULL)

head(hub.result$hub_result)
hub.result$hub_plot$`HALLMARK-APOPTOSIS`
hub.result$hub_plot$`HALLMARK-INFLAMMATORY-RESPONSE`


}

}
