% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{irGSEA.ridgeplot}
\alias{irGSEA.ridgeplot}
\title{Ridge plot
Easy to show the data distribution by ridge plot}
\usage{
irGSEA.ridgeplot(
  object = NULL,
  method = NULL,
  show.geneset = NULL,
  group.by = NULL,
  color.cluster = NULL,
  cluster.levels = NULL,
  ...
)
}
\arguments{
\item{object}{A Seurat after perform \code{\link{irGSEA.score}}}

\item{method}{A character. It should be one of the followling : AUCell,
UCell, singscore, ssgsea.}

\item{show.geneset}{A character. It should be one of the rownames of
enrichment score matrix.}

\item{group.by}{Default ident when it is set to NULL. You can specify other
column of metadata.}

\item{color.cluster}{A vector. Default "ggsci::pal_igv()(the number of colnames
of enrichment score matrix)" when it is set to NULL.}

\item{cluster.levels}{A vector equal to the number of clusters.}

\item{...}{More parameters pass to \code{\link[ggridges]{geom_density_ridges}}}
}
\value{
ridge plot
}
\description{
Ridge plot
Easy to show the data distribution by ridge plot
}
\examples{

\dontrun{
# load PBMC dataset by R package SeuratData
library(Seurat)
library(SeuratData)
# download 3k PBMCs from 10X Genomics
InstallData("pbmc3k")
data("pbmc3k.final")
pbmc3k.final <- SeuratObject::UpdateSeuratObject(pbmc3k.final)

# Seurat object
pbmc3k.final <- irGSEA.score(object = pbmc3k.final, assay = "RNA",
slot = "data", msigdb = T, species = "Homo sapiens",
category = "H", geneid = "symbol",
method = c("AUCell", "UCell", "singscore", "ssgsea"), kcdf = 'Gaussian')

irGSEA.ridgeplot.plot1 <- irGSEA.ridgeplot(object = pbmc3k.final,
method = "UCell", show.geneset = "HALLMARK-INFLAMMATORY-RESPONSE")
irGSEA.ridgeplot.plot2 <- irGSEA.ridgeplot(object = pbmc3k.final,
method = "ssgsea", show.geneset = "HALLMARK-IL6-JAK-STAT3-SIGNALING")

}

}
