% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irGSEA.vlnplot.R
\name{irGSEA.vlnplot}
\alias{irGSEA.vlnplot}
\title{Half vlnplot}
\usage{
irGSEA.vlnplot(
  object = NULL,
  method = NULL,
  show.geneset = NULL,
  group.by = NULL,
  color.cluster = NULL,
  cluster.levels = NULL
)
}
\arguments{
\item{object}{A Seurat after perform \code{\link{irGSEA.score}}}

\item{method}{A character. It should be one or more of the followling : AUCell,
UCell, singscore, ssgsea, JASMINE, VAM, scSE, VISION, gficf, GSVA, zscore,
plage, wmean, wsum, mdt, viper,  GSVApy, AddModuleScore, pagoda2.}

\item{show.geneset}{A character. It should be one of the rownames of
enrichment score matrix.}

\item{group.by}{Default ident when it is set to NULL. You can specify other
column of metadata.}

\item{color.cluster}{A vector. Default "ggsci::pal_igv()(the number of colnames
of enrichment score matrix)" when it is set to NULL.}

\item{cluster.levels}{A vector equal to the number of clusters.}
}
\value{
vlnplot
}
\description{
Easy to show the data distribution by half vlnplot
}
\examples{
\dontrun{
# load PBMC dataset by R package SeuratData
library(Seurat)
library(SeuratData)
# download 3k PBMCs from 10X Genomics
InstallData("pbmc3k")
data("pbmc3k.final")
pbmc3k.final <- SeuratObject::UpdateSeuratObject(pbmc3k.final)

# Seurat object
pbmc3k.final <- irGSEA.score(object = pbmc3k.final, assay = "RNA",
slot = "data", msigdb = T, species = "Homo sapiens",
category = "H", geneid = "symbol",
method = c("AUCell", "UCell", "singscore", "ssgsea"), kcdf = 'Gaussian')

irGSEA.vlnplot.plot1 <- irGSEA.vlnplot(object = pbmc3k.final,
method = c("AUCell", "UCell", "singscore", "ssgsea"),
show.geneset = "HALLMARK-INFLAMMATORY-RESPONSE")
irGSEA.vlnplot.plot2 <- irGSEA.vlnplot(object = pbmc3k.final,
method = c("AUCell", "UCell", "singscore", "ssgsea"),
show.geneset = "HALLMARK-IL6-JAK-STAT3-SIGNALING")


}

}
