% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{irGSEA.score}
\alias{irGSEA.score}
\title{Calculate enrichment scores from scRNA-seq data}
\usage{
irGSEA.score(
  object = NULL,
  assay = NULL,
  slot = "data",
  min.cells = 3,
  min.feature = 0,
  seeds = 123,
  ncores = 4,
  custom = F,
  geneset = NULL,
  msigdb = T,
  species = "Homo sapiens",
  category = "H",
  subcategory = NULL,
  geneid = "symbol",
  method = c("AUCell", "UCell", "singscore", "ssgsea"),
  aucell.MaxRank = NULL,
  ucell.MaxRank = NULL,
  kcdf = "Gaussian"
)
}
\arguments{
\item{object}{Seurat object or scRNA-seq matrix}

\item{assay}{Name of assay to use, defaults to the active assay. The parameter
works when a seurat object is input.}

\item{slot}{Default data. The parameter works if a seurat object is input.}

\item{min.cells}{The minimum detected cells per gene, default 3. The parameter
worsk if a scRNA-seq matrix is input.}

\item{min.feature}{The minimum genes per cell, default 0. The parameter works
if a scRNA-seq matrix is input.}

\item{seeds}{Default 123}

\item{ncores}{Default 4}

\item{custom}{Default False. Set it to true when input own genesets.}

\item{geneset}{Default NULL. Input own genesets as a list. Each element in
the list is a gene set. You can also specify positive and negative genes by
adding a + or - sign in special gene set.}

\item{msigdb}{Default True. You can acquire the collection gene sets from
msigdb database. It will be ignored when custom is set to true.}

\item{species}{Default Homo sapiens. Use msigdbr::msigdbr_show_species() to view
all available species. The parameter works if msigdb is True.}

\item{category}{Default H. You can acquire the Hallmarker gene sets. Use
msigdbr::msigdbr_collections to view all available collections gene sets. The
parameter works if msigdb is True.}

\item{subcategory}{Default NULL. The parameter works if msigdb is True.}

\item{geneid}{Default symbol. Other options are "entrez" and "ensembl". The
parameter works if msigdb is True.}

\item{method}{A vector. Default all method.}

\item{aucell.MaxRank}{The threshold to calculate the AUC. Default only the top 5%
of the expressed genes are used to checks whether the gene set are within
the top 5% when it set to NULL. You can inputc special number, such as 1500,
 to change the threshold. The parameter works if "AUCell" is selected in "method".}

\item{ucell.MaxRank}{Maximum number of genes to rank per cell. Above this
rank, a given gene is considered as not expressed. Default 1500 when it set
to NULL. You can input special number, such as 2000, to increase the rank
range. The parameter works if "UCell" is selected in "method".}

\item{kcdf}{Default Gaussian if input expression values are continuous in
logarithmic scale. Other option is "Poisson" if input expression values are
integer counts. The parameter works if "ssgsea" is selected in "method".}
}
\value{
a Seurat object including score matrix.
}
\description{
Input a Seurat object or scRNA-seq matrix, calculate the enrichment scores of
AUCell, UCell, singscore and ssgsea. Then, return a Seurat object including these
score matrix.
}
\examples{
\dontrun{
# load PBMC dataset by R package SeuratData
library(Seurat)
library(SeuratData)
# download 3k PBMCs from 10X Genomics
InstallData("pbmc3k")
data("pbmc3k.final")
pbmc3k.final <- SeuratObject::UpdateSeuratObject(pbmc3k.final)

# Seurat object
pbmc3k.final <- irGSEA.score(object = pbmc3k.final, assay = "RNA", slot = "data",
msigdb = T, species = "Homo sapiens", category = "H", geneid = "symbol",
method = c("AUCell", "UCell", "singscore", "ssgsea"), kcdf = 'Gaussian')

# data.fram, matrix, or dgmatrix
pbmc3k.final2 <- irGSEA.score(object = pbmc3k.final@assays$RNA@counts,
assay = "RNA", slot = "data", min.cells = 3, min.feature = 0,
method = c("AUCell", "UCell", "singscore", "ssgsea"), kcdf = 'Poisson')

# custom geneset
markers <- list()
markers$Tcell_gd <- c("TRDC+", "TRGC1+", "TRGC2+", "TRDV1+","TRAC-","TRBC1-","TRBC2-")
markers$Tcell_NK <- c("FGFBP2+", "SPON2+", "KLRF1+", "FCGR3A+", "CD3E-","CD3G-")
markers$Tcell_CD4 <- c("CD4","CD40LG")
markers$Tcell_CD8 <- c("CD8A","CD8B")
markers$Tcell_Treg <- c("FOXP3","IL2RA")
pbmc3k.final3 <- irGSEA.score(object = pbmc3k.final, assay = "RNA", slot = "data",
custom = T, geneset = markers, method = c("AUCell", "UCell", "singscore", "ssgsea"),
kcdf = 'Gaussian')
}

}
