% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irGSEA.upset.R
\name{irGSEA.upset}
\alias{irGSEA.upset}
\title{Upset plot}
\usage{
irGSEA.upset(
  object = NULL,
  method = "RRA",
  upset.width = 13,
  upset.height = 7,
  title.size = 10,
  text.size = 9,
  cluster.color = NULL,
  bar.color = "black",
  cluster.levels = NULL,
  mode = "distinct",
  set.size = 1,
  set.degree = NULL,
  table.generate = F,
  ...
)
}
\arguments{
\item{object}{A list after perform \code{\link{irGSEA.integrate}}}

\item{method}{A character. It should be one of the followling : AUCell, UCell,
singscore, ssgsea or RRA. Default RRA.}

\item{upset.width}{Width of the whole upset plot. Default 13.}

\item{upset.height}{Height of the whole upset plot. Default 7.}

\item{title.size}{The fointsize of rownames. Default 10.}

\item{text.size}{The fointsize of rownames. Default 9.}

\item{cluster.color}{A vector. Default "ggsci::pal_igv()(the number of colnames
of enrichment score matrix)" when it is set to NULL.}

\item{bar.color}{A character. Default "black" when it is set to NULL.}

\item{cluster.levels}{A vector equal to the number of clusters.}

\item{mode}{A character. It should be one of the followling : distinct,
intersect, or union. Default distinct. It represents the mode for forming
the combination set. see Mode section \url{https://jokergoo.github.io/ComplexHeatmap-reference/book/upset-plot.html}
for details.}

\item{set.size}{The minimal combination set size. Default 1.}

\item{set.degree}{A vector. Show all combination sets when it set to NULL.
It would show different combination set when it is set to different number.
For example, it only show the interactions among two cluster or three cluster
when it's set to 2 or 3.}

\item{table.generate}{Deault FALSE. It will output a list including all
combination sets and their gene sets when it set to TRUE.}

\item{...}{More parameters pass to \code{\link[ComplexHeatmap]{UpSet}}}
}
\value{
upset plot or list
}
\description{
Easy to show analysis results by upset plot
}
\examples{
\dontrun{
# load PBMC dataset by R package SeuratData
library(Seurat)
library(SeuratData)
# download 3k PBMCs from 10X Genomics
InstallData("pbmc3k")
data("pbmc3k.final")
pbmc3k.final <- SeuratObject::UpdateSeuratObject(pbmc3k.final)

# Seurat object
pbmc3k.final <- irGSEA.score(object = pbmc3k.final, assay = "RNA",
slot = "data", msigdb = T, species = "Homo sapiens",
category = "H", geneid = "symbol",
method = c("AUCell", "UCell", "singscore", "ssgsea"), kcdf = 'Gaussian')

# Integrated analysis
result.dge <- irGSEA.integrate(object = pbmc3k.final,
group.by = "seurat_annotations", metadata = NULL, col.name = NULL,
method = c("AUCell","UCell","singscore","ssgsea"))

irGSEA.upset.plot1 <- irGSEA.upset(object = result.dge, method = "RRA")
irGSEA.upset.plot2 <- irGSEA.upset(object = result.dge, method = "ssgsea")

}


}
